sys/log/common: Package for common log API
This package exposes parts of the log API that are shared by both the
full and stub log implementations.
