+++ sys/log/common/pkg.yml
@@ -6,7 +6,7 @@
 # to you under the Apache License, Version 2.0 (the
 # "License"); you may not use this file except in compliance
 # with the License.  You may obtain a copy of the License at
-# 
+#
 #  http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing,
@@ -17,34 +17,9 @@
 # under the License.
 #
 
-pkg.name: sys/log/full
-pkg.description: Logging utilities for embedded systems.
+pkg.name: sys/log/common
+pkg.description: API common to both the full and stub log implementations.
 pkg.author: "Apache Mynewt <dev@mynewt.apache.org>"
 pkg.homepage: "http://mynewt.apache.org/"
 pkg.keywords:
     - logging
-
-pkg.deps:
-    - kernel/os
-    - sys/flash_map
-    - sys/log/common
-    - util/cbmem
-pkg.deps.LOG_FCB:
-    - hw/hal
-    - fs/fcb
-pkg.deps.LOG_CLI:
-    - sys/shell
-
-pkg.deps.LOG_NEWTMGR:
-    - mgmt/mgmt
-    - encoding/cborattr
-    - encoding/tinycbor
-
-pkg.apis:
-    - log
-
-pkg.init:
-    log_init: 100
-
-pkg.init.LOG_FCB_SLOT1:
-    log_init_slot1: 101
