+++ sys/log/stub/include/log/log.h
@@ -21,7 +21,7 @@
 
 #include <inttypes.h>
 #include "os/mynewt.h"
-#include "log/ignore.h"
+#include "log_common/log_common.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -32,28 +32,6 @@ extern "C" {
 #define LOG_WARN(__l, __mod, ...) IGNORE(__VA_ARGS__)
 #define LOG_ERROR(__l, __mod, ...) IGNORE(__VA_ARGS__)
 #define LOG_CRITICAL(__l, __mod, ...) IGNORE(__VA_ARGS__)
-
-#define LOG_LEVEL_DEBUG         0
-#define LOG_LEVEL_INFO          1
-#define LOG_LEVEL_WARN          2
-#define LOG_LEVEL_ERROR         3
-#define LOG_LEVEL_CRITICAL      4
-/* Up to 7 custom log levels. */
-#define LOG_LEVEL_MAX           UINT8_MAX
-#define LOG_SYSLEVEL            UINT8_MAX
-
-/* Logging medium */
-#define LOG_STORE_CONSOLE    1
-#define LOG_STORE_CBMEM      2
-#define LOG_STORE_FCB        3
-
-/* Global log info */
-struct log_info {
-    uint32_t li_next_index;
-    uint8_t li_version;
-};
-
-struct log_info g_log_info;
 
 struct log {
 };
